requirements.txt
python-novaclient==3.4.0
boto3==1.3.0
paramiko==1.16.0
tornado==4.3
